% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_fb15k.R
\name{fb15k_folder}
\alias{fb15k_folder}
\title{set up folder with fb15k demo data}
\usage{
fb15k_folder(exdir = tempdir())
}
\arguments{
\item{exdir}{character(1) path to destination}
}
\value{
vector of full paths to demo data
}
\description{
set up folder with fb15k demo data
}
\note{
vector returned is named with 'train', 'test', 'valid' for downstream ordering
}
\examples{
tfold = fb15k_folder()
}
