% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini_from_edgeR.R
\name{gini}
\alias{gini}
\title{gini coefficients from rows of a matrix}
\usage{
gini(x)
}
\arguments{
\item{x}{matrix}
}
\value{
vector of gini coefficients
}
\description{
gini coefficients from rows of a matrix
}
\note{
source extracted from edgeR 'gini'
}
