% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_config.R
\name{make_pbg_config}
\alias{make_pbg_config}
\title{produce configuration torchbiggraph fun, with defaults for FB15k test}
\usage{
make_pbg_config(
  entity_path = "data/FB15k",
  edge_paths = list("data/FB15k/freebase_mtr100_mte100-train_partitioned",
    "data/FB15k/freebase_mtr100_mte100-valid_partitioned",
    "data/FB15k/freebase_mtr100_mte100-test_partitioned"),
  checkpoint_path = "model/fb15k",
  entities = list(all = list(num_partitions = 1L)),
  relations = list(list(name = "all_edges", lhs = "all", rhs = "all", operator =
    "complex_diagonal")),
  dynamic_relations = TRUE,
  dimension = 400L,
  global_emb = FALSE,
  comparator = "dot",
  num_epochs = 50L,
  batch_size = 5000L,
  num_uniform_negs = 1000L,
  loss_fn = "softmax",
  lr = 0.1,
  regularization_coef = 0.001,
  eval_fraction = 0,
  num_gpus = 1L
)
}
\arguments{
\item{entity_path}{character(1) path to entity text}
}
\description{
produce configuration torchbiggraph fun, with defaults for FB15k test
}
\examples{
pbg = try(reticulate::import("torchbiggraph"))
if (!inherits(pbg, "try-error")) {
  confdict = make_pbg_config()
}
}
