% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges_to_hdf5.R
\name{make_rel_schema}
\alias{make_rel_schema}
\title{produce RelationSchema for torchbiggraph configuration}
\usage{
make_rel_schema(
  name = "all_edges",
  lhs = "all",
  rhs = "all",
  weight = 1,
  operator = "complex_diagonal",
  all_negs = FALSE,
  pbgref
)
}
\arguments{
\item{name}{character(1)}

\item{lhs}{character(1)}

\item{rhs}{character(1)}

\item{weight}{numeric(1)}

\item{operator}{character(1)}

\item{all_negs}{character(1)}

\item{pbgref}{python reference to torchbiggraph module}
}
\description{
produce RelationSchema for torchbiggraph configuration
}
