% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_fb15k.R
\name{path_to_15k_tgz}
\alias{path_to_15k_tgz}
\title{cache and/or retrieve path to fb15k.tgz}
\usage{
path_to_15k_tgz(cache = BiocFileCache::BiocFileCache())
}
\description{
cache and/or retrieve path to fb15k.tgz
}
