% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimreds.R
\name{pca_CG}
\alias{pca_CG}
\title{update a SingleCellExperiment with PCA (via irlba) reductions of 
embeddings produced with PyTorch-BigGraph}
\usage{
pca_CG(trout, sce, ncomp = 3, ...)
}
\arguments{
\item{trout}{output of `BiocPBG::train_eval`}

\item{sce}{SingleCellExperiment instance derived from or consistent with
the one from which `trout` was produced}

\item{ncomp}{numeric(1) passed as `n` to `irlba::prcomp_irlba`}

\item{\dots}{named arguments other than `n` to be passed to irlba}
}
\value{
The `sce` is updated to include a `reducedDim` component named `PCA`
based on the cell embedding, and the
rowData component is updated to include the PCA applied to gene
embedding
}
\description{
update a SingleCellExperiment with PCA (via irlba) reductions of 
embeddings produced with PyTorch-BigGraph
}
\examples{
data(nn50)
data(t3k)
twp = pca_CG(nn50, t3k)
pairs(reducedDim(twp)$PCA, pch=19, col=factor(twp$celltype3))
}
