% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{t3k}
\alias{t3k}
\title{a SingleCellExperiment for PBMC3K that includes assay elements
logcounts (produced by scran::logNormCounts) and disc (produced
by BiocPBG::disc_matrix)}
\format{
SingleCellExperiment
}
\usage{
t3k
}
\description{
a SingleCellExperiment for PBMC3K that includes assay elements
logcounts (produced by scran::logNormCounts) and disc (produced
by BiocPBG::disc_matrix)
}
\examples{
data(t3k)
assayNames(t3k)
}
\keyword{datasets}
