% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_eval.R
\name{train_eval}
\alias{train_eval}
\title{run training and evaluation}
\usage{
train_eval(tripout, pbg)
}
\arguments{
\item{tripout}{list resulting from `triples_to_hdf5`}

\item{pbg}{torchbiggraph reference}
}
\description{
run training and evaluation
}
\examples{
example(triples_to_hdf5)  # creates tt with tt$config
pyatt = import("attr")
tt2 = list()
tt2$config = pyatt$evolve(tt$config, num_epochs=4L, num_uniform_negs=1000L,
   num_gpus=1L)
train_eval(tt2, pbg)
}
