% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges_to_hdf5.R
\name{triples_to_hdf5}
\alias{triples_to_hdf5}
\title{produce HDF5 files for edges defined in tsv files of triples (left, rel, right)}
\usage{
triples_to_hdf5(
  triple_paths,
  num_partitions = 1L,
  relname = "all_edges",
  lhs = "all",
  rhs = "all",
  weight = 1,
  operator = "complex_diagonal",
  all_negs = FALSE,
  pbgref,
  pathlibref
)
}
\arguments{
\item{triple_paths}{character(3) with test, train, valid edge sets in order}

\item{num_partitions}{integer(1)}

\item{relname}{character(1), passed to make_rel_schema}

\item{lhs}{character(1), passed to make_rel_schema}

\item{rhs}{character(1), passed to make_rel_schema}

\item{weight}{numeric(1), passed to make_rel_schema}

\item{operator}{character(1), passed to make_rel_schema}

\item{all_negs}{character(1), passed to make_rel_schema}

\item{pbgref}{python reference to torchbiggraph module}

\item{pathlibref}{python reference to pathlib module}
}
\description{
produce HDF5 files for edges defined in tsv files of triples (left, rel, right)
}
\examples{
tfis = fb15k_folder()
pbg = reticulate::import("torchbiggraph")
palib = reticulate::import("pathlib")
nn = triples_to_hdf5(tfis, pbgref = pbg, pathlibref=palib)
nn
rhdf5::h5ls(dir(nn$edge_paths[[1]], full=TRUE))
}
