% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimreds.R
\name{umap_CG}
\alias{umap_CG}
\title{update a SingleCellExperiment with UMAP reductions of 
embeddings produced with PyTorch-BigGraph}
\usage{
umap_CG(trout, sce, ncomp = 3, ...)
}
\arguments{
\item{trout}{output of `BiocPBG::train_eval`}

\item{sce}{SingleCellExperiment instance derived from or consistent with
the one from which `trout` was produced}

\item{ncomp}{numeric(1) passed as `n_components` to `uwot::umap`}

\item{\dots}{named arguments other than `n` to be passed to `uwot::umap`}
}
\value{
The `sce` is updated to include a `reducedDim` component named `UMAP`
based on the cell embedding, and the
rowData component is updated to include the UMAP applied to gene
embedding
}
\description{
update a SingleCellExperiment with UMAP reductions of 
embeddings produced with PyTorch-BigGraph
}
