% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_config.R
\name{make_pbg_config}
\alias{make_pbg_config}
\title{produce configuration torchbiggraph fun, with defaults for FB15k test}
\usage{
make_pbg_config(
  entity_path = "data/FB15k",
  edge_paths = list("data/FB15k/freebase_mtr100_mte100-train_partitioned",
    "data/FB15k/freebase_mtr100_mte100-valid_partitioned",
    "data/FB15k/freebase_mtr100_mte100-test_partitioned"),
  checkpoint_path = "model/fb15k",
  entities = list(all = list(num_partitions = 1L)),
  relations = list(list(name = "all_edges", lhs = "all", rhs = "all", operator =
    "complex_diagonal")),
  dynamic_relations = TRUE,
  dimension = 400L,
  global_emb = FALSE,
  comparator = "dot",
  num_epochs = 50L,
  batch_size = 5000L,
  num_uniform_negs = 1000L,
  loss_fn = "softmax",
  lr = 0.1,
  regularization_coef = 0.001,
  eval_fraction = 0,
  num_gpus = 1L
)
}
\arguments{
\item{entity_path}{character(1) path to entity text}
}
\description{
produce configuration torchbiggraph fun, with defaults for FB15k test
}
\examples{
fbpaths = fb15k_folder() # .txt files
ppaths = gsub(".txt", "_partitioned", fbpaths)
fbpl = lapply(fbpaths, force)
pppl = lapply(ppaths, force) # vec2list
pbg = try(reticulate::import("torchbiggraph"))
palib = reticulate::import("pathlib")
txtpl = reticulate::r_to_py(lapply(fbpaths, function(x) palib$Path(x)))
if (!inherits(pbg, "try-error")) {
  confdict = make_pbg_config(entity_path = dirname(fbpl[[1]]),
                               edge_paths=txtpl)
  pbg$converters$importers$convert_input_data( reticulate::dict(all=pbg$config$EntitySchema(num_partitions=1L)),
      list(confdict$relations),
      palib$Path(confdict$entity_path), confdict$edge_paths,
      pppl, # input edge paths
      pbg$converters$importers$TSVEdgelistReader(lhs_col=0L, rhs_col=2L, rel_col=1L),
      dynamic_relations = confdict$dynamic_relations)
}
}
