% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nn50}
\alias{nn50}
\title{embeddings for PBMC3K based on 50 epochs of training by
torch-biggraph for 4780 highly variable genes on 2700 cells.}
\format{
named list
}
\usage{
nn50
}
\description{
embeddings for PBMC3K based on 50 epochs of training by
torch-biggraph for 4780 highly variable genes on 2700 cells.
}
\note{
Built using sce_to_embeddings with t3k
}
\examples{
data(nn50)
names(nn50)
nn50$call
}
\keyword{datasets}
